moved generator logic to seperate file, removed buttonwidget and generator code from app js
