+++ debug.log
@@ -1,0 +1,2 @@
+[0410/090353.199:ERROR:crash_report_database_win.cc(426)] unexpected header
+[0411/085123.539:ERROR:crash_report_database_win.cc(426)] unexpected header
