+++ package.json
@@ -1,25 +1,26 @@
 {
-    "name": "coffeeidle",
-    "version": "0.1.0",
-    "private": true,
-    "homepage": "https://vidsneezes.github.io/CoffeeIdle/",
-    "devDependencies": {
-        "gh-pages": "^0.12.0",
-        "react-scripts": "0.9.0"
-    },
-    "dependencies": {
-        "immutability-helper": "^2.1.2",
-        "react": "^15.4.2",
-        "react-dom": "^15.4.2",
-        "react-redux": "^5.0.4",
-        "redux": "^3.6.0"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test --env=jsdom",
-        "eject": "react-scripts eject",
-        "predeploy": "npm run build",
-        "deploy": "gh-pages -d build"
-    }
+  "name": "coffeeidle",
+  "version": "0.1.0",
+  "private": true,
+  "homepage": "https://vidsneezes.github.io/CoffeeIdle/",
+  "devDependencies": {
+    "gh-pages": "^0.12.0",
+    "react-scripts": "0.9.0"
+  },
+  "dependencies": {
+    "immutability-helper": "^2.1.2",
+    "immutable": "^3.8.1",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-redux": "^5.0.4",
+    "redux": "^3.6.0"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build"
+  }
 }
