+++ src/App.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import update from 'immutability-helper';
 import logo from './logo.svg';
 import './App.css';
+import ButtonWidget from './components/buttons/ButtonWidget.jsx';
 import idleData from './CoffeeData.json';
 
 function arrayToDict(obj,arr){
@@ -16,41 +17,6 @@ function calculateCost(baseCost,rate,owned){
 
 function calculateProduction(generator){
   return generator.productionBase * generator.amount;
-}
-
-class ButtonWidget extends Component{
-  render(){
-    return(
-      <button className={this.props.style} onClick={()=>this.props.onClick()}>
-      {this.props.text}
-      </button>
-    );
-  }
-}
-
-class Generator extends Component {
-  render(){
-    let buttonStatus = "";
-    if(this.props.disabled){
-      buttonStatus = "-dis";
-    }
-
-    return(
-      <section className="generator-container">
-        <div className="generatorcontent-right">
-          <button className="rectangle" onClick={()=>this.props.onClick()} disabled={this.props.disabled}>
-           Buy
-          </button>
-        </div>
-        <div className="generatorcontent-left">
-          <h5 className="center">Owned:{this.props.amount}</h5>
-          <p className="center">{this.props.text}</p>
-          <p className="center">{"costs: "+this.props.cost}</p>
-          <p className="center">{this.props.production+"$/s"}</p>
-        </div>
-      </section>
-    );
-  }
 }
 
 class ClickArea extends Component{
